fix results

+++ console/migrations/m170130_071528_create_athletes_classes_table.php
@@ -15,7 +15,7 @@ class m170130_071528_create_athletes_classes_table extends Migration
 		$this->createTable('AthletesClasses', [
 			'id'          => $this->primaryKey(),
 			'title'       => $this->string()->notNull(),
-			'percent'     => $this->integer()->notNull(),
+			'percent'     => $this->double()->notNull(),
 			'sort'        => $this->integer()->notNull(),
 			'description' => $this->text()
 		]);
